Update README.md
Removed the Meta-discourse at the end of the sentence.
Removed the conjunction "however" to produce a better formal sentence and prevent forced cohesion.
Rewrote some of the sentences to improve flow and "fluff" words. 
Re-wrote the first sentence to remove the passive voice and turned it into an active voice.